libraries/microsoft-agents-a365-notifications/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "tzdata"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-agents-a365-notifications"
dynamic = ["version"]
authors = [
    { name = "Microsoft", email = "support@microsoft.com" },
]
description = "Notification and messaging extensions for AI agent applications"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
license = {text = "MIT"}
keywords = ["notifications", "messaging", "agents", "ai", "routing", "sdk"]
dependencies = [
    "typing-extensions >= 4.0.0",
    "microsoft-agents-activity>=0.4.0",
    "microsoft-agents-hosting-core>=0.4.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/microsoft/Agent365-python"
Repository = "https://github.com/microsoft/Agent365-python"
Issues = "https://github.com/microsoft/Agent365-python/issues"
Documentation = "https://github.com/microsoft/Agent365-python/tree/main/libraries/microsoft-agents-a365-notifications"

[project.optional-dependencies]
dev = [
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "ruff >= 0.1.0",
    "black >= 23.0.0",
    "mypy >= 1.0.0",
]
test = [
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools]
license-files = ["../../LICENSE"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["../../LICENSE"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint.flake8-copyright]
notice-rgx = "# Copyright \\(c\\) Microsoft Corporation\\.\\r?\\n# Licensed under the MIT License\\."
min-file-size = 1

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
